latest changes

+++ TBS_Test/Assets/Scripts/Units/GameUnit.cs
@@ -521,7 +521,7 @@ public class GameUnit : Photon.MonoBehaviour, ISelectable {
 			this.texNetDirSpriteDL = (string)stream.ReceiveNext();
 			this.texNetDirSpriteUR = (string)stream.ReceiveNext();
 			this.texNetDirSpriteDR = (string)stream.ReceiveNext();
-			this.eGridDirection = (UNIT_DIR)stream.ReceiveNext();
+			this.eGridDirection = (UNIT_DIR)((int)stream.ReceiveNext());
 		}
 	}
